Support ASN.1 indefinite length.
Indefinite length is not part of DER and therefore we'd rather not support
it at all. However some quirky certificates use it, so make it possible to
parse them.

This only supports indef length encoding for Sequence, SequenceOf and SetOf
types, as well as outer indef encoding for explicit tags. String types,
Choice and Any are not supported yet - hopefully we'll never see that in
practice.

Additionally, this fixes some unit test bugs and sets strict=False when
printing certificates from the utility.

Thanks to @rep for initially reporting the issue and the test certificate
used for testing.