Update SQLiteDB<>::tree_size_ lazily.
Every time that we do an insertion or a lookup that is contiguous to
tree_size_, increment it accordingly. Calling TreeSize() will scan
forward, but doing it lazily instead of at startup avoids a scan of the
database: the first thing we do is build a tree, by scanning from the
first entry until it gets a lookup error, which will have the same
effect.

The downside is that we can't trust tree_size_ to be fully advanced to
the contiguous tree size, so TreeSize() has to scan, which makes it
blocking. This isn't as bad as it seems, as it needs to take the lock,
which can be held by other threads doing blocking operations, so it was
effectively already blocking.