Don't specify bogus value for evhtp_CFLAGS
Based on feedback from @pphaneuf on #1062.

It seems that libevhtp 1.2.10 installs headers into
$INSTALL_DIR/include; I originally got the $INSTALL_DIR/include/evhtp
directory from the value that 1.2.11's pkg-config file returns, but the
install location has changed in between the two versions.

Given that $INSTALL_DIR/include is in the global CPPFLAGS, we don't need
to specify this again for evhtp_CFLAGS, but to make configure happy we
probably have to pass /something/ in, so pass in the empty string.

When we upgrade to evhtp 1.2.11 we can remove these parameters and
evhtp's pkg-config will mean this stuff will be automagically found by
the configure script.