Tweak signature verification for latest BoringSSL
This commit contains two changes:

1.) The latest BoringSSL does not allow signatureAlgorithm parameters
with large universal tag numbers. (Such certificates fail to parse as of
https://bugs.chromium.org/p/boringssl/issues/detail?id=18.) One
of the test certificates in cert_test.cc had a large universal tag
number, so this commit changes the tests to a.) test that the
certificate doesn't parse with latest BoringSSL, and b.) use a different
certificate (one with a parseable but mismatched signatureAlgorithm
parameter) to test mismatched signature algorithms.

2.) The latest version of BoringSSL switched around error codes for
signature verification. This commit puts the latest version of BoringSSL
into the non-boring OpenSSL code path for handling errors to verify
signatures, and uses ERR_R_EVP_LIB to distinguish signature verification
errors, since that's what both non-boring OpenSSL and recent BoringSSL
have in common. This is likely a temporary situation until we fix
https://bugs.chromium.org/p/boringssl/issues/detail?id=18 in BoringSSL.