Hide away the Cert ctor from ScopedX509.
The hidden version now requires the ScopedX509 passed in to be non-null,
and a FromX509 factory is added that can handle getting null (returning
a null unique_ptr<Cert> accordingly), for those call sites that depended
on the previous looser behaviour.

With this commit, the x509_ member (which is now const) is now
guaranteed to be non-null at all time.