Fix RootFromInclusionProof
RootFromInclusionProof make the assumption that leafs are indexed from
1. However, upon observation from real output from CT server, leafs
appear to be indexed from 0. After inserting a single entry into an
empty tree, the tree size is 1, and the leaf index is 0.

This CL adds a test to the Go Merkle Tree Verifier for this case and
adjusts the verification algorithm to more closely match the python
implementation.

To support testing, this PR also includes the functions needed to
serialize a MerkleTreeLeaf for X509 leaf entries.