[go] Update library code with Go 1.5 version
This adds the Certificate.UnhandledCriticalExtensions field, and
moves policing of this from ParseCertificate() to Certificate.Verify().

As a result, any code that was previously relying on CT's laxer
behaviour for unhandled critical extensions in ParseCertificate(),
may now need to be updated to clear out the
Certificate.UnhandledCriticalExtensions field before calling
Certificate.Verify().