Remove "invalid-url" test from jsonclient
That URL is actually not invalid, plus the test breaks the IPv6 compliance
tests I'm working on.

I tried coming up with an really invalid URL, but it's harder than it seems
(libs tend to fix things silently, by adding escapes, for example). In the end I
decided the test doesn't add much value, as it's attempting to test net/http and
not the client logic itself, so I removed it instead.